✨ feat(tenon-flow): update dev flow

+++ packages/tenon-materials/scripts/build.ts
@@ -3,6 +3,7 @@ import execa from 'execa';
 import fs from "fs";
 import { PhaseName, setPhase } from '@tenon/flow';
 import { debounce } from 'lodash';
+import chokidar from 'chokidar';
 
 const typingsPath = path.resolve(__dirname, '../typings');
 const distPath = path.resolve(__dirname, '../dist');
@@ -13,15 +14,21 @@ const client = createClient();
 let firstBoot = true;
 
 function build() {
-  fs.rmdirSync(cachePath, {
-    recursive: true,
-  });
-  fs.rmdirSync(typingsPath, {
-    recursive: true,
-  });
-  fs.rmdirSync(distPath, {
-    recursive: true,
-  });
+  if (fs.existsSync(cachePath)) {
+    fs.rmdirSync(cachePath, {
+      recursive: true,
+    });
+  }
+  if (fs.existsSync(typingsPath)) {
+    fs.rmdirSync(typingsPath, {
+      recursive: true,
+    });
+  }
+  if (fs.existsSync(distPath)) {
+    fs.rmdirSync(distPath, {
+      recursive: true,
+    });
+  }
 
   const b = Date.now();
   execa.command('tsc', {
@@ -38,18 +45,15 @@ function build() {
 }
 
 
-fs.watch(
-  path.resolve(__dirname, '../components'),
-  {
-    persistent: true,
-    recursive: true,
-  },
-  debounce((event, fileName) => {
-    if (fileName) {
-      console.log(`\n检测到${fileName}变更，即将重新构建物料`);
-      build();
-    }
-  }, 1000),
-);
+const handler = debounce((event, fileName) => {
+  if (fileName) {
+    console.log(`\n检测到${fileName}变更，即将重新构建物料`);
+    build();
+  }
+}, 1000);
 
-build();
+chokidar.watch(path.resolve(__dirname, '../components')).on('all', (event, path) => {
+  handler(event, path);
+});
+
+// build();
